position-only arguments, explicit top-level exports, and put array wrapping in its own module
